moved calculateAccuracy/Consistency to calculations folder
